Reimplemented HUD face with play-scope animation
An infinitely easier solution: instead of creating a fake state machine in ui scope, I simply spawn a non-interactive invisible actor for each player, attach pointers to the player to said actor, and animate it in play scope. Then the HUD gets access to said actor (via an event handler), and uses GetSpriteTexture to convert it to a TextureID and draws it on the screen. It just works (tm)!