Differentiated move brake application in FPP and TPP
In first person: your horizontal speed is simply reduced (in all directions) if you've landed and aren't moving, OR if you're pressing movement keys in the direction that is opposite to your current vel. (This is how this originally worked.)

In TPP: at ALL times, if you press movement keys in the direction opposite to your velocity vector, you'll change direction immediately (snappy turns, non-slippery landing).